+++ frontend/src/feature/posts/types/dtos/createCommentRequestDto.ts
@@ -1,0 +1,4 @@
+export type createCommentRequestDto = {
+  postId: string;
+  content: string;
+}
